# openapi.yaml
openapi: 3.0.3
info:
  title: Euchre Endpoint
  description: >-
    This is the remote interface for invocation of Java-based algorithms from
    the `euchre-plt` tournament framework (see
    [euchre-plt](https://github.com/crashka/euchre-plt)).  The server side only
    plays hands for individual games; the client side (which we will call the
    "coordinator") manages the larger context of deals and accounting for matches,
    tournaments, etc.
  license:
    name: MIT License
    url: https://opensource.org/license/mit
  version: '0.5'
...
paths:
  /session:
    post:
      tags:
        - Session
      summary: Start a new session
      description: >-
        Notification from the coordinator that a new session has been started; the
        session is identified by a token that will be included in all requests and
        responses for the session
      operationId: postSession
      requestBody:
        description: |-
          `token` - Session identifier (same for all calls within the session)\
          `status` - Must have a value of "**new**"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionInfo'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; session created locally\
            `token` - *Copied from request*\
            `status` - Should be set to "**active**"\
            `cards` *[optional\*]* - Card representation - int[24]: 9C, 10C, JC, ...
            9D, 10D, JD, ... QS, KS, AS\
            `suits` *[optional\*]* - Suit representation - int[4]: Clubs, Diamonds,
            Hearts, Spades\
            \
            \* Server may optionally specify local representations for `cards` and 
            and `suits` (null value indicates server default, same as [0..23] and
            [0..3] respectively)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionProto'
    patch:
      tags:
        - Session
      summary: Update the status of a session
      description: >-
        Notification from the coordinator that the status of the session has
        changed; currently only a status of "**complete**" is supported
      operationId: patchSession
      requestBody:
        description: |-
          `token` - Session identifier\
          `status` - Updated status; valid values:
          - "**complete**" - Session is complete, local resources may be
          released

          `info` - Purely informational string (e.g. scores, stats, etc.)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionComplete'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; session updated (or removed) locally\
            Response body (optional) is an echo of the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionComplete'
  /game:
    post:
      tags:
        - Game
      summary: Start a new game
      description: >-
        Notification from the coordinator that a new game has been started
      operationId: postGame
      requestBody:
        description: |-
          `token` - Session identifier\
          `gameNum` - Sequential game number within the session\
          `status` - Must have a value of "**new**"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameInfo'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; game created locally\
            `token` - *Copied from request*\
            `gameNum` - *Copied from request*\
            `status` - Should be set to "**active**"\
            `info` - Optional (ignored)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatus'
    patch:
      tags:
        - Game
      summary: Update the status of a game
      description: >-
        Notification from the coordinator that the status of the game has
        changed; currently only a status of "**complete**" is supported
      operationId: patchGame
      requestBody:
        description: |-
          `token` - Session identifier\
          `gameNum` - Sequential game number within the session\
          `status` - Updated status; valid values:
          - "**complete**" - Game is complete, local resources may be released

          `info` - Purely informational string (e.g. scores, stats, etc.)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameComplete'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; game updated (or removed) locally\
            Response body (optional) is an echo of the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameComplete'
  /deal:
    post:
      tags:
        - Deal
      summary: Start a new deal
      description: >-
        Notification from the coordinator that a new deal has been started; subsequent
        requests or notifications will be made for individual actions (bids, card plays,
        etc.) within the deal
      operationId: postDeal
      requestBody:
        description: |-
          `token` - Session identifier\
          `gameNum` - Sequential game number within the session (0-n)\
          `dealNum` - Sequential deal number within the game (0-n)\
          `status` - Must have a value of "**new**"\
          `cards` - Cards dealt - int[24]:
          - 0-4 : Position 0 hand
          - 5-9 : Position 1 hand
          - 10-14: Position 2 hand
          - 15-19: Position 3 hand (dealer)
          - 20 : Turn card
          - 21-23: Buries
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealInfo'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; deal created locally\
            `token` - *Copied from request*\
            `gameNum` - *Copied from request*\
            `dealNum` - *Copied from request*\
            `status` - Should be set to "**active**"\
            `info` - Optional (ignored)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealStatus'
    patch:
      tags:
        - Deal
      summary: Update the status of a deal
      description: >-
        Notification from the coordinator that the status of the deal has
        changed; currently only a status of "**complete**" is supported
      operationId: patchDeal
      requestBody:
        description: |-
          `token` - Session identifier\
          `gameNum` - Sequential game number within the session (0-n)\
          `dealNum` - Sequential deal number within the game (0-n)\
          `status` - Updated status; valid values:
          - "**complete**" - Deal is complete, local resources may be released

          `info` - Purely informational string (e.g. scores, stats, etc.)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealComplete'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; deal updated (or removed) locally\
            Response body (optional) is an echo of the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealComplete'
  /bid:
    get:
      tags:
        - Bid
      summary: Request bid from server
      description: >-
        Note that multiple bids per round may be requested, for partners hosted on the
        same server
      operationId: getBid
      parameters:
        - name: token
          in: query
          required: true
          description: Session identifier
          schema:
            $ref: '#/components/schemas/SessionToken'
        - name: gameNum
          in: query
          required: true
          description: Game number (0-n)
          schema:
            type: integer
            example: 10
        - name: dealNum
          in: query
          required: true
          description: Deal number within the game (0-n)
          schema:
            type: integer
            example: 3
        - name: round
          in: query
          required: true
          description: Bidding round (0-1)
          schema:
            type: integer
            example: 0
        - name: turnCard
          in: query
          required: true
          description: >-
            Turn card (0-23, indicates biddable suit) [round 0], or -1 [round 1]
          schema:
            type: integer
            example: 14
        - name: pos
          in: query
          required: true
          description: Bidding position (0-3, with 3 = dealer)
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: |-
            Successful operation; bid made\
            `token` - *Copied from request*\
            `gameNum` - *Copied from request*\
            `dealNum` - *Copied from request*\
            `round` - *Copied from request*\
            `turnCard` - *Copied from request*\
            `pos` - *Copied from request*\
            `suit` - Suit declaration (0-3), or -1 = pass\
            `alone` - Going alone declaration (true/false)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidInfo'
    post:
      tags:
        - Bid
      summary: Notify server of a bid
      description: >-
        Notification from the coordinator of a remote bid
      operationId: postBid
      requestBody:
        description: |-
          `token` - Session identifier\
          `gameNum` - Sequential game number within the session (0-n)\
          `dealNum` - Sequential deal number within the game (0-n)\
          `round` - Bidding round (0-1)\
          `turnCard` - Turn card (0-23) [round 0], or -1 [round 1]\
          `pos` - Bidding position (0-3, with 3 = dealer)\
          `suit` - Suit declaration (0-3), or -1 = pass\
          `alone` - Going alone declaration (true/false)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BidInfo'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; bid recorded locally\
            `token` - *Copied from request*\
            `gameNum` - *Copied from request*\
            `dealNum` - *Copied from request*\
            `round` - *Copied from request*\
            `turnCard` - *Copied from request*\
            `pos` - *Copied from request*\
            `suit` *[optional\*]* - ***Suggested*** suit declaration (0-3), or -1
            = pass; or null\
            `alone` *[optional\*]* - ***Suggested*** going alone declaration
            (true/false); or null\
            \
            \* Server may optionally specify suggested values for `suit` and `alone`
            (null value indicates no suggestion)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidSugg'
  /swap:
    get:
      tags:
        - Swap
      summary: Request swap from server
      description: >-
        Only called if server represents the dealer position for a first round bid
        declaration (optionally called if partner declares a loner)
      operationId: getSwap
      parameters:
        - name: token
          in: query
          required: true
          description: Session identifier
          schema:
            $ref: '#/components/schemas/SessionToken'
        - name: gameNum
          in: query
          required: true
          description: Game number (0-n)
          schema:
            type: integer
            example: 10
        - name: dealNum
          in: query
          required: true
          description: Deal number (0-n) within the game
          schema:
            type: integer
            example: 3
        - name: declarerPos
          in: query
          required: true
          description: Declarer position (0-3, with 3 = dealer)
          schema:
            type: integer
            example: 1
        - name: turnCard
          in: query
          required: true
          description: Turn card (0-23, indicates trump suit)
          schema:
            type: integer
            example: 14
        - name: pos
          in: query
          required: true
          description: >-
            Swapping position (always 3, denoting dealer) *[included for
            completeness]*
          schema:
            type: integer
            example: 3
        - name: swappableCards
          in: query
          required: true
          description: Swappable cards - int[5] (0-23)
          schema:
            type: array
            items:
              type: integer
            example: [2, 8, 11, 15, 21]
      responses:
        '200':
          description: |-
            Successful operation; card swap made\
            `token` - *Copied from request*\
            `gameNum` - *Copied from request*\
            `dealNum` - *Copied from request*\
            `declarerPos` - *Copied from request*\
            `turnCard` - *Copied from request*\
            `pos` - *Copied from request*\
            `card` - Card swapped for turn card (0-23), or -1 = no swap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapInfo'
    post:
      tags:
        - Swap
      summary: Notify server of a swap
      description: >-
        Notification from the coordinator of a swap by a remote dealer (note: this
        is optionally posted, depending on whether coordinator wants to share all
        hidden information with server)
      operationId: postSwap
      requestBody:
        description: |-
          `token` - Session identifier\
          `gameNum` - Sequential game number within the session (0-n)\
          `dealNum` - Sequential deal number within the game (0-n)\
          `declarerPos` - Declarer position (0-3, with 3 = dealer)\
          `turnCard` - Turn card (0-23, indicates trump suit)\
          `pos` - Swapping position (always 3, denoting dealer)\
          `card` - Card swapped for turn card (0-23), or -1 = no swap
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapInfo'
        required: true
      responses:
        '200':
          description: |-
            Successful operation; card swap recorded locally\
            `token` - *Copied from request*\
            `gameNum` - *Copied from request*\
            `dealNum` - *Copied from request*\
            `declarerPos` - *Copied from request*\
            `turnCard` - *Copied from request*\
            `pos` - *Copied from request*\
            `card` *[optional\*]* - ***Suggested*** card to swap (0-23), or -1
            = no swap; or null\
            \
            \* Server may optionally specify suggested value for `card` (null value
            indicates no suggestion)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapSugg'
...
components:
  schemas:
    SessionToken:
      type: string
      description: Session identifier
      example: 464f31d
    SessionInfo:
      type: object
      description: Session information
      properties:
        token:
          $ref: '#/components/schemas/SessionToken'
        status:
          type: string
          description: Session status (new)
          example: new
          enum:
            - new
    SessionProto:
      type: object
      description: Session status with server-defined protocol information
      properties:
        token:
          $ref: '#/components/schemas/SessionToken'
        status:
          type: string
          description: Session status (active)
          example: active
          enum:
            - active
        cards:
          type: array
          description: >-
            Card representation - int[24]: 9C, 10C, JC, ... 9D, 10D, JD, ... QS, KS, AS
          items:
            type: integer
          example: [3, 7, 11, 15, 19, 23,
                    2, 6, 10, 14, 18, 22,
                    1, 5,  9, 13, 17, 21,
                    0, 4,  8, 12, 16, 20]
        suits:
          type: array
          description: >-
            Suit representation - int[4]: clubs, diamonds, hearts, spades
          items:
            type: integer
          example: [3, 2, 1, 0]
...
    SessionComplete:
      type: object
      description: Session status (complete)
      properties:
        token:
          $ref: '#/components/schemas/SessionToken'
        status:
          type: string
          description: Session status (complete)
          example: complete
          enum:
            - complete
        info:
          type: string
          description: Session status information
          example: ''
    GameInfo:
      type: object
      description: Game information
      properties:
        token:
          $ref: '#/components/schemas/SessionToken'
        gameNum:
          type: integer
          description: Game number (0-n)
          example: 10
        status:
          type: string
          description: Game status (new)
          example: new
          enum:
            - new
    GameStatus:
      type: object
      description: Game status
      properties:
        token:
          $ref: '#/components/schemas/SessionToken'
        gameNum:
          type: integer
          description: Game number (0-n)
          example: 10
        status:
          type: string
          description: Game status
          example: active
          enum:
            - new
            - active
            - complete
        info:
          type: string
          description: Game status information
          example: ''
    GameComplete:
      type: object
      description: Game status (complete)
      properties:
        token:
          $ref: '#/components/schemas/SessionToken'
        gameNum:
          type: integer
          description: Game number (0-n)
          example: 10
        status:
          type: string
          description: Game status (complete)
          example: complete
          enum:
            - complete
        info:
          type: string
          description: Game status information
          example: ''
    DealInfo:
      type: object
      description: Deal information
      properties:
        token:
          $ref: '#/components/schemas/SessionToken'
        gameNum:
          type: integer
          description: Game number (0-n)
          example: 10
        dealNum:
          type: integer
          description: Deal number (0-n)
          example: 3
        status:
          type: string
          description: Deal status (new)
          example: new
          enum:
            - new
        cards:
          type: array
          description: |-
            Cards dealt - int[24]:
            - 0-4 : Position 0 hand
            - 5-9 : Position 1 hand
            - 10-14: Position 2 hand
            - 15-19: Position 3 hand (dealer)
            - 20 : Turn card
            - 21-23: Buries
          items:
            type: integer
          example: [22,  1, 21, 12,  7,
                    23, 17, 15, 14, 10,
                    20, 16,  3, 13,  4,
                    5,  11,  2,  9,  0,
                    8,  19,  6, 18]
...